+++ web_transaction_test.go
@@ -70,7 +70,7 @@ func TestClient_DeleteWebTransaction(t *testing.T) {
 	})
 
 	var client = &Client{APIEndpoint: server.URL, AuthToken: "foo"}
-	id := 1
+	id := int64(1)
 	err := client.DeleteWebTransaction(id)
 
 	if err != nil {
